Added basic API functionality
