+++ src/app/api/test/route.ts
@@ -1,0 +1,12 @@
+import { sql } from '@vercel/postgres';
+import { NextResponse } from 'next/server';
+
+export async function GET(request: any)
+{
+    try {
+        const result = await sql`SELECT * FROM test WHERE name = 'Ali'`;
+        return NextResponse.json(result.rows);
+    } catch (error: any) {
+        return NextResponse.json({ error: error.message }, { status: 500 });
+    }
+}
