+++ src/app/page.tsx
@@ -1,139 +1,37 @@
-'use client';
-
-import React, { useState, useEffect } from "react";
-import { Calendar, Scheduler, useArrayState } from "@cubedoodl/react-simple-scheduler";
-import { Checkbox, FormGroup, FormControlLabel} from "@mui/material";
+'use client'
+import React, { useState } from "react";
+import Link from "next/link";
+import { TextField } from "@mui/material";
 
 export default function Home(){
-  const [selected, setSelected] = useState(new Date());
 
-  // Master event array to render selected events
-  const [events, setEvents, addEvent] = useArrayState(null);
+    const [username, setUsername] = useState(null);
 
-  // Temporary events to store individual events
-  const [ericEvents, setEricEvents, addEricEvent] = useArrayState(null);
-  const [felicityEvents, setFelicityEvents, addFelicityEvent] = useArrayState(null);
-  const [annieEvents, setAnnieEvents, addAnnieEvent] = useArrayState(null);
-
-  // Dictionary state variable to store booleans dictating whether the corresponding user's events should be displayed
-  const [filterStates, setFilterStates] = useState<{ [key: string]: boolean }>({
-    "Eric": false,
-    "Felicity": false,
-    "Annie": false
-  });
-
-  // Storing the calendar user
-  const [user, setUser] = useState("Eric");
-
-  // Checking if the filter button is checked
-  const [filter, setFilter] = useState(false);
-
-  // Editing the truth values of the filter states
-  const handleClick = (e: React.ChangeEvent<HTMLInputElement>) => {
-      let filterKey = "";
-      if (e.target.id === "EricCheck") {
-          filterKey = "Eric"
-      }
-      if (e.target.id === "FelicityCheck") {
-          filterKey = "Felicity"
-      }
-      if (e.target.id === "AnnieCheck") {
-          filterKey = "Annie"
-      }
-      setFilterStates(prevState => ({
-        ...prevState,
-        [filterKey]: !prevState[filterKey]
-      }));
-  }
-
-  // Event filtering
-  useEffect(() => {
-    console.log(filterStates)
-    let tempEvents: any[] = [];
-    if (filterStates["Eric"]) {
-      tempEvents = tempEvents.concat(ericEvents);
+    // Functions to detect username change and submission
+    const handleUsernameFieldChange = (event: any) => {
+        setUsername(event.target.value);
     }
-    if (filterStates["Felicity"]) {
-      tempEvents = tempEvents.concat(felicityEvents);
+    const handleUsernameSubmit = async (event: any) => {
+        event.preventDefault();
+        const response = await fetch('api/test');
+        const result = await response.json();
+        console.log(result[0].message);
     }
-    if (filterStates["Annie"]) {
-      tempEvents = tempEvents.concat(annieEvents);
-    }
-    setEvents(tempEvents);
-  }, [filterStates])
 
-  
+    // Temporary function to test querying
 
-  return (
-    <>
-    <div className="flex justify-end items-center gap-10">
-      <select
-        value={user}
-        onChange={(e) => setUser(e.target.value)}
-      >
-        <option value="Eric">Eric</option>
-        <option value="Felicity">Felicity</option>
-        <option value="Annie">Annie</option>
-      </select>
-      <button className={`border p-2 rounded-md text-white ${
-        filter ? 'bg-green-500 border-green-500 hover:bg-green-700 hover:bg-green-700' : 'bg-red-500 border-red-500 hover:bg-red--700 hover:border-red-500'
-      }`} id="filter" onClick={() => setFilter(!filter)}>Filter</button>
-      <FormGroup className="h-100">
-        <FormControlLabel control={<Checkbox size="large" id="EricCheck" onChange={handleClick}/>} label="Eric" />
-        <FormControlLabel control={<Checkbox size="large" id="FelicityCheck" onChange={handleClick}/>} label="Felicity" />
-        <FormControlLabel control={<Checkbox size="large" id="AnnieCheck" onChange={handleClick}/>} label="Annie" />
-      </FormGroup>
-      <Scheduler
-              
-              // This is horrendous. I promise to fix it
-              events={events}
-              selected={selected}
-              setSelected={setSelected}
-              onRequestAdd={(evt) => {
-                switch (user) {
-                  case 'Eric':
-                    evt.style={
-                      color: "white",
-                      backgroundColor: "blue"
-                    }
-                    addEricEvent(evt)
-                    break;
-                  case 'Felicity':
-                    evt.style={
-                      color: "white",
-                      backgroundColor: "green"
-                    }
-                    addFelicityEvent(evt)
-                    break;
-                  case 'Annie':
-                  evt.style={
-                    color: "white",
-                    backgroundColor: "red"
-                  }
-                  addAnnieEvent(evt)
-                  break;
-                }
-                addEvent(evt)
-              }}
-              onRequestEdit={(evt) => alert("Edit element requested")}
-              style={{
-                container: {
-                  width: "80%",
-                  height: "100%",
-                  color: "black",
-                  padding: "20px"
-                },
-                head: {
-                  // TODO
-                },
-                body: {
-                  // TODO
-                }
-              }}
-            />
-          
-    </div>
-      
-    </>
-  );
+    return (
+        <>
+            <div className="flex flex-col gap-4 items-center justify-center w-screen h-screen">
+                <Link href="/calendar" className="border border-black p-2 rounded-md w-36 text-center">
+                    Login
+                </Link>
+                <form onSubmit={handleUsernameSubmit} className="flex flex-col gap-2 justify-center items-center">
+                    <TextField onChange={handleUsernameFieldChange} id="outlined-basic" label="Username" variant="outlined"/>
+                    <button type="submit" className="border border-black rounded-md w-24">Submit</button>
+                </form>
+               
+            </div>
+        </>
+    )
 }
